+++ src/Components/Reports/Common/ExecutiveSecondPage.js
@@ -1,97 +1,141 @@
 import React, { Fragment } from 'react';
-import { Text } from '@react-pdf/renderer';
-import { TachometerAltIconConfig } from '@patternfly/react-icons/dist/js/icons/tachometer-alt-icon';
-import { AngleDoubleDownIconConfig } from '@patternfly/react-icons/dist/js/icons/angle-double-down-icon';
-import { AngleDoubleUpIconConfig } from '@patternfly/react-icons/dist/js/icons/angle-double-up-icon';
-import { AutomationIconConfig } from '@patternfly/react-icons/dist/js/icons/automation-icon';
-import { InProgressIconConfig } from '@patternfly/react-icons/dist/js/icons/in-progress-icon';
-import { CheckCircleIconConfig } from '@patternfly/react-icons/dist/js/icons/check-circle-icon';
-import styles from './styles';
-import { DescriptionList } from './DescriptionList';
+import { InstancesTable } from './InstancesTable';
 
 export const ExecutiveSecondPage = () => {
 
-    const stateDescription = [
-        {
-            title: 'Optimized',
-            description: 'Performing at an optimal level',
-            iconPath: CheckCircleIconConfig.svgPath,
-            iconScale: 0.014,
-            fillColor: 'green'
-        },
-        {
-            title: 'Under pressure',
-            description: 'Peaking occasionally',
-            iconPath: TachometerAltIconConfig.svgPath,
-            iconScale: 0.014,
-            fillColor: '#030303'
-        },
-        {
-            title: 'Undersized',
-            description: 'Using more than 80% of system resources',
-            iconPath: AngleDoubleDownIconConfig.svgPath,
-            iconScale: 0.014,
-            fillColor: 'red'
-        },
-        {
-            title: 'Oversized',
-            description: 'Using less than 20% of system resources',
-            iconPath: AngleDoubleUpIconConfig.svgPath,
-            iconScale: 0.014,
-            fillColor: '#f09800'
-        },
-        {
-            title: 'Idling',
-            description: 'Consuming less than 5% of resources',
-            iconPath: AutomationIconConfig.svgPath,
-            iconScale: 0.008,
-            fillColor: '#030303'
-        },
-        {
-            title: 'Waiting for data',
-            description: 'Data has not been received or is being processed. Initial data processing takes up to 24 hours.',
-            iconPath: InProgressIconConfig.svgPath,
-            iconScale: 0.008,
-            fillColor: '#2B9AF3'
-        }
-    ];
+    const currentInstancesDetails = {
+        heading: 'Current instance types',
+        description: 'This is your current state, that represents reality of how your instance types are doing',
+        data: [
+            {
+                type: 't3.small',
+                systemCount: '30',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't4g.small',
+                systemCount: '42',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't1.micro',
+                systemCount: '44',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't3.micro',
+                systemCount: '20',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't3.nano',
+                systemCount: '10',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            }
+        ]
+    };
 
-    const conditionsDescription = [
-        {
-            title: 'CPU pressure',
-            description: 'CPU registered peaks higher than 20% over several one-minute time periods'
-        },
-        {
-            title: 'Disk I/O  pressure',
-            description: 'Disk I/O registered peaks higher than 20% over several one-minute time periods'
-        },
-        {
-            title: 'RAM pressure',
-            description: 'RAM registered peaks higher than 20% over several one-minute time periods'
-        }
-    ];
+    const suggestedInstancesDetails = {
+        heading: 'Most suggested instance types (yesterday)',
+        description: 'Yesterday we identified these instance types',
+        data: [
+            {
+                type: 't3.nano',
+                systemCount: '40',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't4g.nano',
+                systemCount: '20',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't1.micro',
+                systemCount: '30',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't3.micro',
+                systemCount: '20',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't2.nano',
+                systemCount: '10',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            }
+        ]
+    };
 
-    return <Fragment key="second-page">
-        <Text style={styles.execHeading}>Description of states</Text>
-        {
-            stateDescription.map(
-                (state) => <DescriptionList
-                    key={state.title}
-                    title={state.title}
-                    description={state.description}
-                    iconPath={state.iconPath}
-                    iconScale={state.iconScale}
-                    fillColor={state.fillColor} />)
-        }
+    const historicalInstancesDetails = {
+        heading: 'Most suggested instance types (45 days)',
+        description: 'In the last 45 days we suggested you these instances # of times',
+        data: [
+            {
+                type: 't3.nano',
+                systemCount: '30',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't2.nano',
+                systemCount: '42',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't1.micro',
+                systemCount: '44',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 't3.micro',
+                systemCount: '20',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            },
+            {
+                type: 'm1.small',
+                systemCount: '10',
+                // eslint-disable-next-line max-len
+                description: 'Intel Xeon E5-2680 v2 (Ivy Bridge) instance with 8 vCPUs and 15 GiB of RAM, running on AWS eu-west-1 and eu-west-2 regions'
+            }
+        ]
+    };
 
-        <Text style={styles.execHeading}>Description of conditions</Text>
-        {
-            conditionsDescription.map(
-                (condition) => <DescriptionList
-                    key={condition.title}
-                    title={condition.title}
-                    description={condition.description}/>)
-        }
+    return <Fragment key="third-page">
+        <InstancesTable
+            id={'current_instance_types'}
+            instanceDetails={currentInstancesDetails.data}
+            heading={currentInstancesDetails.heading}
+            description={currentInstancesDetails.description}
+
+        />
+
+        <InstancesTable
+            id='suggested_instance_types'
+            instanceDetails={suggestedInstancesDetails.data}
+            heading={suggestedInstancesDetails.heading}
+            description={suggestedInstancesDetails.description}
+        />
+
+        <InstancesTable
+            id='historical_instance_types'
+            instanceDetails={historicalInstancesDetails.data}
+            heading={historicalInstancesDetails.heading}
+            description={historicalInstancesDetails.description}
+        />
     </Fragment>;
 };
